+++ rpc/backend/tx_info_test.go
@@ -19,7 +19,148 @@ import (
 
 	"github.com/zeta-chain/zetacore/rpc/backend/mocks"
 	rpctypes "github.com/zeta-chain/zetacore/rpc/types"
+	"github.com/zeta-chain/zetacore/testutil/sample"
 )
+
+func (suite *BackendTestSuite) TestGetSyntheticTransactionByHash() {
+	hash := sample.Hash().Hex()
+	_, txRes := suite.buildSyntheticTxResult(hash)
+
+	suite.backend.indexer = nil
+	client := suite.backend.clientCtx.Client.(*mocks.Client)
+	query := fmt.Sprintf(
+		"%s.%s='%s'",
+		evmtypes.TypeMsgEthereumTx,
+		evmtypes.AttributeKeyEthereumTxHash,
+		common.HexToHash(hash).Hex(),
+	)
+	queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+	RegisterBaseFee(queryClient, sdk.NewInt(1))
+	RegisterTxSearchWithTxResult(client, query, []byte{}, txRes)
+	RegisterBlock(client, 1, nil)
+	RegisterBlockResultsWithTxResults(client, 1, []*abci.ResponseDeliverTx{&txRes})
+
+	res, err := suite.backend.GetTransactionByHash(common.HexToHash(hash))
+	suite.Require().NoError(err)
+
+	// assert fields
+	suite.Require().Equal(hash, res.Hash.Hex())
+	nonce, _ := hexutil.DecodeUint64(res.Nonce.String())
+	suite.Require().Equal(uint64(1), nonce)
+	suite.Require().Equal(int64(1), res.BlockNumber.ToInt().Int64())
+	suite.Require().Equal("0x775b87ef5D82ca211811C1a02CE0fE0CA3a455d7", res.To.Hex())
+	suite.Require().Equal("0x735b14BB79463307AAcBED86DAf3322B1e6226aB", res.From.Hex())
+	txIndex, _ := hexutil.DecodeUint64(res.TransactionIndex.String())
+	suite.Require().Equal(uint64(8888), txIndex)
+	txType, _ := hexutil.DecodeUint64(res.Type.String())
+	suite.Require().Equal(uint64(88), txType)
+	suite.Require().Equal(int64(7001), res.ChainID.ToInt().Int64())
+	suite.Require().Equal(int64(1000), res.Value.ToInt().Int64())
+	suite.Require().Nil(res.V)
+	suite.Require().Nil(res.R)
+	suite.Require().Nil(res.S)
+}
+
+func (suite *BackendTestSuite) TestGetSyntheticTransactionReceiptByHash() {
+	hash := sample.Hash().Hex()
+	_, txRes := suite.buildSyntheticTxResult(hash)
+
+	suite.backend.indexer = nil
+	client := suite.backend.clientCtx.Client.(*mocks.Client)
+	query := fmt.Sprintf(
+		"%s.%s='%s'",
+		evmtypes.TypeMsgEthereumTx,
+		evmtypes.AttributeKeyEthereumTxHash,
+		common.HexToHash(hash).Hex(),
+	)
+	queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+	var header metadata.MD
+	RegisterParams(queryClient, &header, 1)
+	RegisterParamsWithoutHeader(queryClient, 1)
+	RegisterTxSearchWithTxResult(client, query, []byte{}, txRes)
+	RegisterBlock(client, 1, nil)
+	RegisterBlockResultsWithTxResults(client, 1, []*abci.ResponseDeliverTx{&txRes})
+
+	res, err := suite.backend.GetTransactionReceipt(common.HexToHash(hash))
+	suite.Require().NoError(err)
+
+	// assert fields
+	suite.Require().Equal(common.HexToHash(hash), res["transactionHash"])
+	blockNumber, _ := hexutil.DecodeUint64(res["blockNumber"].(hexutil.Uint64).String())
+	suite.Require().Equal(uint64(1), blockNumber)
+	toAddress := common.HexToAddress("0x775b87ef5D82ca211811C1a02CE0fE0CA3a455d7")
+	fromAddress := common.HexToAddress("0x735b14BB79463307AAcBED86DAf3322B1e6226aB")
+	suite.Require().Equal(&toAddress, res["to"])
+	suite.Require().Equal(fromAddress, res["from"])
+	status, _ := hexutil.DecodeUint64(res["status"].(hexutil.Uint).String())
+	suite.Require().Equal(uint64(1), status)
+	txType, _ := hexutil.DecodeUint64(res["type"].(hexutil.Uint).String())
+	suite.Require().Equal(uint64(88), txType)
+	txIndex, _ := hexutil.DecodeUint64(res["transactionIndex"].(hexutil.Uint64).String())
+	suite.Require().Equal(uint64(8888), txIndex)
+}
+
+func (suite *BackendTestSuite) TestGetSyntheticTransactionByBlockNumberAndIndex() {
+	hash := sample.Hash().Hex()
+	tx, txRes := suite.buildSyntheticTxResult(hash)
+
+	suite.backend.indexer = nil
+	client := suite.backend.clientCtx.Client.(*mocks.Client)
+	queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+	RegisterBlock(client, 1, []types.Tx{tx})
+	RegisterBlockResultsWithTxResults(client, 1, []*abci.ResponseDeliverTx{&txRes})
+	RegisterBaseFee(queryClient, sdk.NewInt(1))
+
+	res, err := suite.backend.GetTransactionByBlockNumberAndIndex(rpctypes.BlockNumber(1), 0)
+	suite.Require().NoError(err)
+
+	// assert fields
+	suite.Require().Equal(hash, res.Hash.Hex())
+	nonce, _ := hexutil.DecodeUint64(res.Nonce.String())
+	suite.Require().Equal(uint64(1), nonce)
+	suite.Require().Equal("0x775b87ef5D82ca211811C1a02CE0fE0CA3a455d7", res.To.Hex())
+	suite.Require().Equal("0x735b14BB79463307AAcBED86DAf3322B1e6226aB", res.From.Hex())
+	txType, _ := hexutil.DecodeUint64(res.Type.String())
+	suite.Require().Equal(uint64(88), txType)
+	suite.Require().Equal(int64(7001), res.ChainID.ToInt().Int64())
+	suite.Require().Equal(int64(1000), res.Value.ToInt().Int64())
+	suite.Require().Nil(res.V)
+	suite.Require().Nil(res.R)
+	suite.Require().Nil(res.S)
+}
+
+func (suite *BackendTestSuite) TestGetSyntheticTransactionByBlockNumberAndIndexWithRealTransaction() {
+	hash := sample.Hash().Hex()
+	tx, txRes := suite.buildSyntheticTxResult(hash)
+	msgEthereumTx, _ := suite.buildEthereumTx()
+
+	realTx := suite.signAndEncodeEthTx(msgEthereumTx)
+
+	suite.backend.indexer = nil
+	client := suite.backend.clientCtx.Client.(*mocks.Client)
+	queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
+	// synthetic tx with real MsgEthereumTx
+	RegisterBlock(client, 1, []types.Tx{realTx, tx})
+	RegisterBlockResultsWithTxResults(client, 1, []*abci.ResponseDeliverTx{{}, &txRes})
+	RegisterBaseFee(queryClient, sdk.NewInt(1))
+
+	res, err := suite.backend.GetTransactionByBlockNumberAndIndex(rpctypes.BlockNumber(1), 1)
+	suite.Require().NoError(err)
+
+	// assert fields
+	suite.Require().Equal(hash, res.Hash.Hex())
+	nonce, _ := hexutil.DecodeUint64(res.Nonce.String())
+	suite.Require().Equal(uint64(1), nonce)
+	suite.Require().Equal("0x775b87ef5D82ca211811C1a02CE0fE0CA3a455d7", res.To.Hex())
+	suite.Require().Equal("0x735b14BB79463307AAcBED86DAf3322B1e6226aB", res.From.Hex())
+	txType, _ := hexutil.DecodeUint64(res.Type.String())
+	suite.Require().Equal(uint64(88), txType)
+	suite.Require().Equal(int64(7001), res.ChainID.ToInt().Int64())
+	suite.Require().Equal(int64(1000), res.Value.ToInt().Int64())
+	suite.Require().Nil(res.V)
+	suite.Require().Nil(res.R)
+	suite.Require().Nil(res.S)
+}
 
 func (suite *BackendTestSuite) TestGetTransactionByHash() {
 	msgEthereumTx, _ := suite.buildEthereumTx()
@@ -45,8 +186,8 @@ func (suite *BackendTestSuite) TestGetTransactionByHash() {
 
 	rpcTransaction, err := rpctypes.NewRPCTransaction(
 		msgEthereumTx.AsTransaction(),
-		common.Hash{},
-		0,
+		common.HexToHash("0x1"),
+		1,
 		0,
 		big.NewInt(1),
 		suite.backend.chainID,
@@ -74,7 +215,7 @@ func (suite *BackendTestSuite) TestGetTransactionByHash() {
 			"fail - Block Result error",
 			func() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
-				RegisterBlock(client, 1, txBz)
+				RegisterBlock(client, 1, []types.Tx{txBz})
 				RegisterBlockResultsError(client, 1)
 			},
 			msgEthereumTx,
@@ -86,7 +227,7 @@ func (suite *BackendTestSuite) TestGetTransactionByHash() {
 			func() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterBlock(client, 1, txBz)
+				RegisterBlock(client, 1, []types.Tx{txBz})
 				RegisterBlockResults(client, 1)
 				RegisterBaseFeeError(queryClient)
 			},
@@ -99,7 +240,7 @@ func (suite *BackendTestSuite) TestGetTransactionByHash() {
 			func() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterBlock(client, 1, txBz)
+				RegisterBlock(client, 1, []types.Tx{txBz})
 				RegisterBlockResults(client, 1)
 				RegisterBaseFee(queryClient, sdk.NewInt(1))
 			},
@@ -459,7 +600,7 @@ func (suite *BackendTestSuite) TestGetTransactionByBlockNumberAndIndex() {
 			func() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
-				RegisterBlock(client, 1, bz)
+				RegisterBlock(client, 1, []types.Tx{bz})
 				RegisterBlockResults(client, 1)
 				RegisterBaseFee(queryClient, sdk.NewInt(1))
 			},
@@ -592,7 +733,7 @@ func (suite *BackendTestSuite) TestGetTransactionReceipt() {
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterParams(queryClient, &header, 1)
 				RegisterParamsWithoutHeader(queryClient, 1)
-				RegisterBlock(client, 1, txBz)
+				RegisterBlock(client, 1, []types.Tx{txBz})
 				RegisterBlockResults(client, 1)
 			},
 			msgEthereumTx,